refactor(GameController): App의 게임 컨트롤 역할을 대신 할 클래스 생성
App은 게임을 실행시키는 역할만 하도록 개선
